Revert "feat(random uuid): require user to pass the mac address instead of randomly generating it"
This reverts commit 203e732c4890c27ff5e659e50cdda7c6edef5be0.
