+++ crates/nu-command/src/random/mod.rs
@@ -6,11 +6,6 @@ mod float;
 mod int;
 mod random_;
 mod uuid;
-mod uuid1;
-mod uuid3;
-mod uuid4;
-mod uuid5;
-mod uuid7;
 
 pub use self::binary::SubCommand as RandomBinary;
 pub use self::bool::SubCommand as RandomBool;
@@ -18,11 +13,5 @@ pub use self::chars::SubCommand as RandomChars;
 pub use self::dice::SubCommand as RandomDice;
 pub use self::float::SubCommand as RandomFloat;
 pub use self::int::SubCommand as RandomInt;
-pub use self::uuid::get_namespace_and_name;
 pub use self::uuid::SubCommand as RandomUuid;
-pub use self::uuid1::SubCommand as RandomUuid1;
-pub use self::uuid3::SubCommand as RandomUuid3;
-pub use self::uuid4::SubCommand as RandomUuid4;
-pub use self::uuid5::SubCommand as RandomUuid5;
-pub use self::uuid7::SubCommand as RandomUuid7;
 pub use random_::RandomCommand as Random;
