+++ crates/nu-command/src/random/uuid.rs
@@ -1,6 +1,5 @@
 use nu_engine::command_prelude::*;
-use nu_protocol::{report_parse_warning, ParseWarning};
-use uuid::Uuid;
+use uuid::{Timestamp, Uuid};
 
 #[derive(Clone)]
 pub struct SubCommand;
@@ -14,55 +13,187 @@ impl Command for SubCommand {
         Signature::build("random uuid")
             .category(Category::Random)
             .input_output_types(vec![(Type::Nothing, Type::String)])
+            .named(
+                "version",
+                SyntaxShape::Int,
+                "The UUID version to generate (1, 3, 4, 5, 7). Defaults to 4 if not specified.",
+                Some('v'),
+            )
+            .named(
+                "namespace",
+                SyntaxShape::String,
+                "The namespace for v3 and v5 UUIDs (dns, url, oid, x500). Required for v3 and v5.",
+                Some('n'),
+            )
+            .named(
+                "name",
+                SyntaxShape::String,
+                "The name string for v3 and v5 UUIDs. Required for v3 and v5.",
+                Some('s'),
+            )
+            .named(
+                "mac",
+                SyntaxShape::String,
+                "The MAC address (node ID) used to generate v1 UUIDs. Required for v1.",
+                Some('m'),
+            )
             .allow_variants_without_examples(true)
     }
 
     fn description(&self) -> &str {
-        "Generate a random uuid4 string."
+        "Generate a random uuid string of the specified version."
     }
 
     fn search_terms(&self) -> Vec<&str> {
-        vec!["generate", "uuid4"]
+        vec!["generate", "uuid4", "uuid1", "uuid3", "uuid5", "uuid7"]
     }
 
     fn run(
         &self,
         engine_state: &EngineState,
-        _stack: &mut Stack,
+        stack: &mut Stack,
         call: &Call,
         _input: PipelineData,
     ) -> Result<PipelineData, ShellError> {
-        let head = call.head;
-        report_parse_warning(
-            &StateWorkingSet::new(engine_state),
-            &ParseWarning::DeprecatedWarning {
-                old_command: "random uuid".into(),
-                new_suggestion: "use `random uuid[version]`".into(),
-                span: head,
-                url: "`help random uuid[version]`".into(),
-            },
-        );
-        uuid(call)
+        uuid(engine_state, stack, call)
     }
 
     fn examples(&self) -> Vec<Example> {
-        vec![Example {
-            description: "Generate a random uuid4 string",
-            example: "random uuid",
-            result: None,
-        }]
+        vec![
+            Example {
+                description: "Generate a random uuid v4 string (default)",
+                example: "random uuid",
+                result: None,
+            },
+            Example {
+                description: "Generate a uuid v1 string (timestamp-based)",
+                example: "random uuid -v 1 -m 00:11:22:33:44:55",
+                result: None,
+            },
+            Example {
+                description: "Generate a uuid v3 string (namespace with MD5)",
+                example: "random uuid -v 3 -n dns -s example.com",
+                result: None,
+            },
+            Example {
+                description: "Generate a uuid v4 string (random).",
+                example: "random uuid -v 4",
+                result: None,
+            },
+            Example {
+                description: "Generate a uuid v5 string (namespace with SHA1)",
+                example: "random uuid -v 5 -n dns -s example.com",
+                result: None,
+            },
+            Example {
+                description: "Generate a uuid v7 string (timestamp + random)",
+                example: "random uuid -v 7",
+                result: None,
+            },
+        ]
     }
 }
 
-fn uuid(call: &Call) -> Result<PipelineData, ShellError> {
+fn uuid(
+    engine_state: &EngineState,
+    stack: &mut Stack,
+    call: &Call,
+) -> Result<PipelineData, ShellError> {
     let span = call.head;
-    let uuid_4 = Uuid::new_v4().hyphenated().to_string();
 
-    Ok(PipelineData::Value(Value::string(uuid_4, span), None))
+    let version: Option<i64> = call.get_flag(engine_state, stack, "version")?;
+    let version = version.unwrap_or(4);
+
+    let uuid_str = match version {
+        1 => {
+            let ts = Timestamp::now(uuid::timestamp::context::NoContext);
+            let node_id = get_mac_address(engine_state, stack, call, span)?;
+            let uuid = Uuid::new_v1(ts, &node_id);
+            uuid.hyphenated().to_string()
+        }
+        3 => {
+            let (namespace, name) = get_namespace_and_name(engine_state, stack, call, span)?;
+            let uuid = Uuid::new_v3(&namespace, name.as_bytes());
+            uuid.hyphenated().to_string()
+        }
+        4 => {
+            let uuid = Uuid::new_v4();
+            uuid.hyphenated().to_string()
+        }
+        5 => {
+            let (namespace, name) = get_namespace_and_name(engine_state, stack, call, span)?;
+            let uuid = Uuid::new_v5(&namespace, name.as_bytes());
+            uuid.hyphenated().to_string()
+        }
+        7 => {
+            let ts = Timestamp::now(uuid::timestamp::context::NoContext);
+            let uuid = Uuid::new_v7(ts);
+            uuid.hyphenated().to_string()
+        }
+        _ => {
+            return Err(ShellError::GenericError {
+                error: format!(
+                    "Unsupported UUID version: {}. Supported versions are 1, 3, 4, 5, and 7.",
+                    version
+                ),
+                msg: "Invalid UUID version".to_string(),
+                span: Some(span),
+                help: Some("Specify version with -v: 1, 3, 4, 5, or 7".to_string()),
+                inner: Vec::new(),
+            });
+        }
+    };
+
+    Ok(PipelineData::Value(Value::string(uuid_str, span), None))
 }
 
-// NOTE: Do not remove this function. It is used in the `uuid3` and `uuid5` modules.
-pub fn get_namespace_and_name(
+fn get_mac_address(
+    engine_state: &EngineState,
+    stack: &mut Stack,
+    call: &Call,
+    span: Span,
+) -> Result<[u8; 6], ShellError> {
+    let mac_str: Option<String> = call.get_flag(engine_state, stack, "mac")?;
+
+    let mac_str = match mac_str {
+        Some(mac) => mac,
+        None => {
+            return Err(ShellError::MissingParameter {
+                param_name: "mac".to_string(),
+                span,
+            });
+        }
+    };
+
+    let mac_parts = mac_str.split(':').collect::<Vec<&str>>();
+    if mac_parts.len() != 6 {
+        return Err(ShellError::IncorrectValue {
+            msg: "MAC address must be in the format XX:XX:XX:XX:XX:XX".to_string(),
+            val_span: span,
+            call_span: span,
+        });
+    }
+
+    let mac: [u8; 6] = mac_parts
+        .iter()
+        .map(|x| u8::from_str_radix(x, 16))
+        .collect::<Result<Vec<u8>, _>>()
+        .map_err(|_| ShellError::IncorrectValue {
+            msg: "MAC address must be in the format XX:XX:XX:XX:XX:XX".to_string(),
+            val_span: span,
+            call_span: span,
+        })?
+        .try_into()
+        .map_err(|_| ShellError::IncorrectValue {
+            msg: "MAC address must be in the format XX:XX:XX:XX:XX:XX".to_string(),
+            val_span: span,
+            call_span: span,
+        })?;
+
+    Ok(mac)
+}
+
+fn get_namespace_and_name(
     engine_state: &EngineState,
     stack: &mut Stack,
     call: &Call,
